pyproject.toml
[tool.poetry]
name = "is-it-down"
version = "0.1.0"
description = ""
authors = ["tvey <liteitt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.1.1"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"
validators = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
